add basic routes navigator
