+++ lib/main.dart
@@ -3,6 +3,7 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+import 'constants/theme.dart';
 import 'package:behomie/features/Home/view/screen.dart';
 
 void main() async {
@@ -19,4 +20,28 @@ void main() async {
       ),
     ),
   );
+}
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      localizationsDelegates: context.localizationDelegates,
+      supportedLocales: context.supportedLocales,
+      locale: context.locale,
+      routes: {
+        '/': (context) => const MyHomePage(),
+      },
+      initialRoute: '/',
+      title: 'BeHomie',
+      debugShowCheckedModeBanner: false,
+      themeMode: ThemeMode.system,
+      theme: ThemeClass
+          .lightTheme, // applies this theme if the device theme is light mode
+      darkTheme: ThemeClass
+          .darkTheme, // applies this theme if the device theme is dark mode
+    );
+  }
 }
